add tests

+++ src/Data/GameDefinition.cs
@@ -5,12 +5,12 @@ namespace Queens.Data;
 internal readonly record struct GameDefinition
 {
     [JsonPropertyName("sideLength")]
-    public ushort SideLength { get; init; }
+    public int SideLength { get; init; }
 
     [JsonPropertyName("colors")]
     public ColorData[] Colors { get; init; }
 
     [JsonPropertyName("cellColors")]
-    public ushort[] CellColors { get; init; }
+    public int[] CellColors { get; init; }
 
 }
